func_80023144 +3 to -1 on one declaration: the narrow-local trap
`c` holds an s8 struct field and I had declared the local s8 to match. Retail
reads the field once with lb and then does sltiu on the same sign-extended
register; the s8 local made gcc read it a SECOND time with lbu and mask with
andi 0xFF to produce (u32)c. Two sites, so lbu +2 and andi +2 -- the whole of
the remaining +3. `s32 c`, the widest natural type with the cast at the point
of use, removes all four and takes the census to a single opcode. Writing
`(u8)c >= 4` on the s8 local instead changes nothing, so it is the
declaration and not the comparison.

What is left is one nop: in two places we fill a load-delay slot retail leaves
empty, both times with an instruction our scheduler had available and retail's
did not.

Build green at 84747e64.

+++ parked/func_80023144.c
@@ -12,7 +12,7 @@ void func_80023144(u8 *arg0, s32 arg1) {
     s32 v;
     s32 mode;
     s32 id;
-    s8 c;
+    s32 c;
     s32 k;
     s32 fl;
     u8 side;
